+++ test-suite/TestVariables/Program.cs
@@ -3,6 +3,7 @@ using System.IO;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Threading;
+using System.Threading.Tasks;
 
 using DbgTest;
 using DbgTest.DAP;
@@ -253,6 +254,49 @@ public struct TestStruct7
     }
 }
 
+//// START .NET 8.0+ block
+public class TestPrimaryConstructorVariables(string primaryText, TestImplicitCast1 primaryObject)
+{
+    public void BreakInInstanceMethod()
+    {
+        int local = primaryText.Length + primaryObject.data;
+        local++;                                            Label.Breakpoint("bp_primary_ctor");
+    }
+
+    public void BreakWithLocalShadow()
+    {
+        string primaryText = "local shadow";
+        int local = primaryText.Length + primaryObject.data;
+        local++;                                            Label.Breakpoint("bp_primary_ctor_local_shadow");
+    }
+
+    public void BreakWithArgumentShadow(string primaryText)
+    {
+        int local = primaryText.Length + primaryObject.data;
+        local++;                                            Label.Breakpoint("bp_primary_ctor_arg_shadow");
+    }
+
+    public void BreakWithLambdaShadow()
+    {
+        string primaryText = "lambda shadow";
+        Func<int> getLength = () => {
+            int local = primaryText.Length + primaryObject.data;
+            local++;                                        Label.Breakpoint("bp_primary_ctor_lambda_shadow");
+            return local;
+        };
+        int local = getLength();
+    }
+
+    public async Task BreakWithAsyncShadow()
+    {
+        string primaryText = "async shadow";
+        await Task.Yield();
+        int local = primaryText.Length + primaryObject.data;
+        local++;                                            Label.Breakpoint("bp_primary_ctor_async_shadow");
+    }
+}
+//// END .NET 8.0+ block
+
 class Program
 {
     static void Main(string[] args)
@@ -273,6 +317,13 @@ class Program
                 Context.AddBreakpoint(@"__FILE__:__LINE__", "bp_func1");
                 Context.AddBreakpoint(@"__FILE__:__LINE__", "bp_func2");
                 Context.AddBreakpoint(@"__FILE__:__LINE__", "bp_getter");
+//// START .NET 8.0+ block
+                Context.AddBreakpoint(@"__FILE__:__LINE__", "bp_primary_ctor");
+                Context.AddBreakpoint(@"__FILE__:__LINE__", "bp_primary_ctor_local_shadow");
+                Context.AddBreakpoint(@"__FILE__:__LINE__", "bp_primary_ctor_arg_shadow");
+                Context.AddBreakpoint(@"__FILE__:__LINE__", "bp_primary_ctor_lambda_shadow");
+                Context.AddBreakpoint(@"__FILE__:__LINE__", "bp_primary_ctor_async_shadow");
+//// END .NET 8.0+ block
                 Context.SetBreakpoints(@"__FILE__:__LINE__");
                 Context.ConfigurationDone(@"__FILE__:__LINE__");
 
@@ -975,6 +1026,99 @@ class Program
 
         TestFunctionArgs(10, 5f, "test_string");
 
+//// START .NET 8.0+ block
+        var primaryConstructorVariables = new TestPrimaryConstructorVariables("primary text", new TestImplicitCast1(321));
+        primaryConstructorVariables.BreakInInstanceMethod();
+        primaryConstructorVariables.BreakWithLocalShadow();
+        primaryConstructorVariables.BreakWithArgumentShadow("argument shadow");
+        primaryConstructorVariables.BreakWithLambdaShadow();
+        primaryConstructorVariables.BreakWithAsyncShadow().GetAwaiter().GetResult();
+
+        Label.Checkpoint("test_primary_constructor", "test_primary_constructor_local_shadow",
+            (Object context) =>
+            {
+                Context Context = (Context)context;
+                Context.WasBreakpointHit(@"__FILE__:__LINE__", "bp_primary_ctor");
+
+                Int64 frameId = Context.DetectFrameId(@"__FILE__:__LINE__", "bp_primary_ctor");
+                int variablesReference = Context.GetVariablesReference(@"__FILE__:__LINE__", frameId, "Locals");
+
+                Context.CheckVariableNameCount(@"__FILE__:__LINE__", variablesReference, "primaryText", 1);
+                Context.EvalVariable(@"__FILE__:__LINE__", variablesReference, "string", "primaryText", "\"primary text\"");
+                Context.GetAndCheckValue(@"__FILE__:__LINE__", frameId, "primaryText", "\"primary text\"");
+                Context.GetAndCheckValue(@"__FILE__:__LINE__", frameId, "primaryObject.data", "321");
+
+                int primaryObjectReference = Context.GetChildVariablesReference(@"__FILE__:__LINE__", variablesReference, "primaryObject");
+                Context.EvalVariable(@"__FILE__:__LINE__", primaryObjectReference, "int", "data", "321");
+
+                Context.Continue(@"__FILE__:__LINE__");
+            });
+
+        Label.Checkpoint("test_primary_constructor_local_shadow", "test_primary_constructor_arg_shadow",
+            (Object context) =>
+            {
+                Context Context = (Context)context;
+                Context.WasBreakpointHit(@"__FILE__:__LINE__", "bp_primary_ctor_local_shadow");
+                Int64 frameId = Context.DetectFrameId(@"__FILE__:__LINE__", "bp_primary_ctor_local_shadow");
+                int variablesReference = Context.GetVariablesReference(@"__FILE__:__LINE__", frameId, "Locals");
+
+                Context.CheckVariableNameCount(@"__FILE__:__LINE__", variablesReference, "primaryText", 1);
+                Context.EvalVariable(@"__FILE__:__LINE__", variablesReference, "string", "primaryText", "\"local shadow\"");
+                Context.GetAndCheckValue(@"__FILE__:__LINE__", frameId, "primaryText", "\"local shadow\"");
+                Context.GetAndCheckValue(@"__FILE__:__LINE__", frameId, "primaryObject.data", "321");
+
+                Context.Continue(@"__FILE__:__LINE__");
+            });
+
+        Label.Checkpoint("test_primary_constructor_arg_shadow", "test_primary_constructor_lambda_shadow",
+            (Object context) =>
+            {
+                Context Context = (Context)context;
+                Context.WasBreakpointHit(@"__FILE__:__LINE__", "bp_primary_ctor_arg_shadow");
+                Int64 frameId = Context.DetectFrameId(@"__FILE__:__LINE__", "bp_primary_ctor_arg_shadow");
+                int variablesReference = Context.GetVariablesReference(@"__FILE__:__LINE__", frameId, "Locals");
+
+                Context.CheckVariableNameCount(@"__FILE__:__LINE__", variablesReference, "primaryText", 1);
+                Context.EvalVariable(@"__FILE__:__LINE__", variablesReference, "string", "primaryText", "\"argument shadow\"");
+                Context.GetAndCheckValue(@"__FILE__:__LINE__", frameId, "primaryText", "\"argument shadow\"");
+                Context.GetAndCheckValue(@"__FILE__:__LINE__", frameId, "primaryObject.data", "321");
+
+                Context.Continue(@"__FILE__:__LINE__");
+            });
+
+        Label.Checkpoint("test_primary_constructor_lambda_shadow", "test_primary_constructor_async_shadow",
+            (Object context) =>
+            {
+                Context Context = (Context)context;
+                Context.WasBreakpointHit(@"__FILE__:__LINE__", "bp_primary_ctor_lambda_shadow");
+                Int64 frameId = Context.DetectFrameId(@"__FILE__:__LINE__", "bp_primary_ctor_lambda_shadow");
+                int variablesReference = Context.GetVariablesReference(@"__FILE__:__LINE__", frameId, "Locals");
+
+                Context.CheckVariableNameCount(@"__FILE__:__LINE__", variablesReference, "primaryText", 1);
+                Context.EvalVariable(@"__FILE__:__LINE__", variablesReference, "string", "primaryText", "\"lambda shadow\"");
+                Context.GetAndCheckValue(@"__FILE__:__LINE__", frameId, "primaryText", "\"lambda shadow\"");
+                Context.GetAndCheckValue(@"__FILE__:__LINE__", frameId, "primaryObject.data", "321");
+
+                Context.Continue(@"__FILE__:__LINE__");
+            });
+
+        Label.Checkpoint("test_primary_constructor_async_shadow", "test_debugger_browsable_state",
+            (Object context) =>
+            {
+                Context Context = (Context)context;
+                Context.WasBreakpointHit(@"__FILE__:__LINE__", "bp_primary_ctor_async_shadow");
+                Int64 frameId = Context.DetectFrameId(@"__FILE__:__LINE__", "bp_primary_ctor_async_shadow");
+                int variablesReference = Context.GetVariablesReference(@"__FILE__:__LINE__", frameId, "Locals");
+
+                Context.CheckVariableNameCount(@"__FILE__:__LINE__", variablesReference, "primaryText", 1);
+                Context.EvalVariable(@"__FILE__:__LINE__", variablesReference, "string", "primaryText", "\"async shadow\"");
+                Context.GetAndCheckValue(@"__FILE__:__LINE__", frameId, "primaryText", "\"async shadow\"");
+                Context.GetAndCheckValue(@"__FILE__:__LINE__", frameId, "primaryObject.data", "321");
+
+                Context.Continue(@"__FILE__:__LINE__");
+            });
+//// END .NET 8.0+ block
+
         TestStruct4 ts4 = new TestStruct4();
 
         int i = 0;
@@ -1128,7 +1272,7 @@ class Program
 
         dummy1 = 2;                                         Label.Breakpoint("bp_func2");
 
-        Label.Checkpoint("bp_func_test2", "test_debugger_browsable_state",
+        Label.Checkpoint("bp_func_test2", "test_primary_constructor",
             (Object context) =>
             {
                 Context Context = (Context)context;